django/requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1;python_version<"3.9"
Baker==1.3
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
coverage==7.3.1
cryptography==39.0.0
defusedxml==0.7.1
Django==4.1.5
django-filter==22.1
django-seed==0.3.1
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.21.4
Faker==16.3.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
model-bakery==1.9.0
oauthlib==3.2.2
packaging==23.0
Pillow==9.4.0
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
python-decouple==3.7
python3-openid==3.2.0
pytz==2022.7
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
testresources==2.0.1
toposort==1.8
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==1.26.13
whitenoise==6.3.0
